+++ lib/sidekiq.rb
@@ -23,6 +23,7 @@ module Sidekiq
     :require => '.',
     :environment => nil,
     :timeout => 8,
+    :profile => false,
   }
 
   def self.options